% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{clean_bea}
\alias{clean_bea}
\title{\code{clean_bea()} downloads WRDS data.}
\usage{
clean_bea(data, start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{data}{Data downloaded from \link{get_bea}}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{clean_bea()} downloads WRDS data.
}
