% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{clean_compustat}
\alias{clean_compustat}
\title{\code{clean_compustat()} cleans Compustat data.}
\usage{
clean_compustat(data, start_date = "1940-01-01", end_date = "2023-12-31")
}
\arguments{
\item{data}{Data downloaded from \link{get_compustat}}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{clean_compustat()} cleans Compustat data.
}
