% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{clean_crsp}
\alias{clean_crsp}
\title{\code{clean_crsp()} downloads WRDS data.}
\usage{
clean_crsp(data, start_date = "1974-01-01", end_date = "2023-12-01", rf = NULL)
}
\arguments{
\item{data}{Data downloaded from \link{get_crsp}}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{rf}{Risk-free rate to compute excess returns. Defaults to \code{NULL}, in which
case excess returns are not computed.}
}
\description{
\code{clean_crsp()} downloads WRDS data.
}
