% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famafrench.R
\name{clean_famafrench}
\alias{clean_famafrench}
\title{Cleans data downloaded from Kenneth French's data library
\code{clean_famafrench()} cleans data downloaded with \link{get_famafrench}.}
\usage{
clean_famafrench(data, start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{data}{Data in a format returned by \code{get_famafrench()}.}

\item{start_date}{Date to start the data at.}

\item{end_date}{Date to end the data at.}
}
\description{
Cleans data downloaded from Kenneth French's data library
\code{clean_famafrench()} cleans data downloaded with \link{get_famafrench}.
}
