% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{clean_fred}
\alias{clean_fred}
\title{\code{clean_fred()} cleans FRED data.}
\usage{
clean_fred(data, start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{data}{Data downloaded from \link{get_fred}}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{clean_fred()} cleans FRED data.
}
