% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{clean_wrds}
\alias{clean_wrds}
\title{\code{get_wrds()} downloads WRDS data.}
\usage{
clean_wrds(data, start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{data}{Data downloaded from \link{get_wrds}}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_wrds()} downloads WRDS data.
}
