% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{get_bea}
\alias{get_bea}
\title{Download data the Bureau of Economic Analysis (BEA)}
\usage{
get_bea(start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_bea()} downloads bea data.
}
