% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{get_compustat}
\alias{get_compustat}
\title{Download data from Compustat using Wharton's WRDS}
\usage{
get_compustat(start_date = "1940-01-01", end_date = "2023-12-31")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_compustat()} downloads data from Compustat
}
