% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{get_crsp}
\alias{get_crsp}
\title{Download data from CRSP using Wharton's WRDS}
\usage{
get_crsp(start_date = "1940-01-01", end_date = "2023-12-31")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_crsp()} downloads data from CRSP
}
