% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp_compustat_merge.R
\name{get_crsp_compustat}
\alias{get_crsp_compustat}
\title{Download linking table from CRSP-Compustat
Merged table and merge CRSP and Compustat}
\usage{
get_crsp_compustat(
  crsp_monthly,
  start_date = "1974-01-01",
  end_date = "2023-12-01"
)
}
\arguments{
\item{crsp_monthly}{A tibble with CRSP monthly data}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_crsp_compustat()} dowloads CRSP-Compustat linking table
}
