% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{get_crsp_daily}
\alias{get_crsp_daily}
\title{\code{get_crsp_daily()} downloads daily frequency data from CRSP}
\usage{
get_crsp_daily(
  start_date = "1974-01-01",
  end_date = "2023-12-01",
  rf = NULL,
  permnos = NULL,
  batch_size = 500
)
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{rf}{Risk-free rate at daily frequency to compute excess returns.
Defaults to \code{NULL}, in which case excess returns are not computed.}

\item{permnos}{Vector of permnos to download. Defaults to \code{NULL}, in which case all
permnos are downloaded}

\item{batch_size}{How many observations to download at a time. Defaults to 500.}
}
\description{
\code{get_crsp_daily()} downloads daily frequency data from CRSP
}
