% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{get_crsp_indices_daily}
\alias{get_crsp_indices_daily}
\title{\code{get_crsp_indices_daily()} downloads daily frequency stock market indices from CRSP}
\usage{
get_crsp_indices_daily(start_date = "1974-01-01", end_date = "2023-01-01")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_crsp_indices_daily()} downloads daily frequency stock market indices from CRSP
}
