% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famafrench.R
\name{get_famafrench}
\alias{get_famafrench}
\title{Download data from Kenneth French's data library}
\usage{
get_famafrench(
  dataset_names = c("Fama/French 3 Factors", "Fama/French 3 Factors [Daily]"),
  ...
)
}
\arguments{
\item{dataset_names}{Input vector with names of datasets to download. Either a character vector, or something
coercible to one.}

\item{...}{Additional arguments passed to \link[frenchdata:download_french_data]{frenchdata::download_french_data}.}
}
\description{
\code{get_famafrench()} downloads Fama-French 3 Factors.
}
