% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_fred}
\alias{get_fred}
\title{Download data from the Federal Reserve Bank of St. Louis FRED}
\usage{
get_fred(fred_series = c("GDPC1", "GDPPOT"), ...)
}
\arguments{
\item{fred_series}{A single character string, a character vector or tibble
representing a single (or multiple) FRED series}

\item{...}{Additional arguments passed to \link[tidyquant:tq_get]{tidyquant::tq_get}.}
}
\description{
\code{get_fred()} downloads FRED data.
}
