% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{get_sic}
\alias{get_sic}
\title{\code{get_sic()} downloads link table between gvkey and sic}
\usage{
get_sic(start_date = "1974-01-01", end_date = "2023-01-01")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_sic()} downloads link table between gvkey and sic
}
