% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds.R
\name{get_wrds}
\alias{get_wrds}
\title{Download data from Wharton's WRDS}
\usage{
get_wrds(start_date = "1950-01-01", end_date = "2023-01-01")
}
\arguments{
\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_wrds()} downloads WRDS data.
}
