% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investmentVolData-package.R
\docType{package}
\name{investmentVolData-package}
\alias{investmentVolData}
\alias{investmentVolData-package}
\title{investmentVolData: Create Data for Paper "Aggregate Investment and Stock Returns"}
\description{
Create Data for Paper "Aggregate Investment and Stock Returns".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernando-duarte/investmentVolData}
  \item Report bugs at \url{https://github.com/fernando-duarte/investmentVolData/issues}
}

}
\author{
\strong{Maintainer}: Fernando Duarte \email{fernando.duarte@gmail.com} (\href{https://orcid.org/0000-0002-8495-6936}{ORCID})

}
\keyword{internal}
