% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap_matA}
\alias{overlap_matA}
\title{\code{overlap_matA()} transformation matrix for regression with overlapping observations}
\usage{
overlap_matA(T, k)
}
\arguments{
\item{T}{Number of observations in the regression}

\item{k}{Number of periods over which returns are computed for the dependent variable}
}
\description{
\code{overlap_matA()} transformation matrix for regression with overlapping observations
}
\references{
\href{https://doi.org/10.1111/j.1468-5957.2011.02244.x}{Britten-Jones, Neuberger and Nolte (2011)}
}
\seealso{
\code{\link[=transformX]{transformX()}}
}
