% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fred_data.R
\name{pull_fred_data}
\alias{pull_fred_data}
\title{Pull data from FRED}
\usage{
pull_fred_data(fred_api_key, replicate = FALSE, vintage_date = NULL)
}
\arguments{
\item{fred_api_key}{API key from FRED}

\item{replicate}{Boolean, use original paper vintage}

\item{vintage_date}{String date ("yyyy-mm-dd"), use given date as vintage}
}
\value{
A tibble/data.frame
}
\description{
Pull data from FRED
}
