% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{ytd_to_q}
\alias{ytd_to_q}
\title{\code{ytd_to_q()} constructs quarterly flow from year-to-date flow taking into
account different fiscal year ends and missing observations}
\usage{
ytd_to_q(data, ytd_var, q_var_name, fiscal_q = "fiscal_q")
}
\arguments{
\item{data}{A tibble of four observations for a fiscal year of one firm}

\item{ytd_var}{The year-to-date flow variable}

\item{q_var_name}{The name of the new quarterly flow variable created}

\item{fiscal_q}{The fiscal quarter of the firm (with values 1,2,3 or 4)}
}
\description{
\code{ytd_to_q()} constructs quarterly flow from year-to-date flow taking into
account different fiscal year ends and missing observations
}
