% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea.R
\name{get_bea}
\alias{get_bea}
\title{Download data the Bureau of Economic Analysis (BEA)}
\usage{
get_bea(
  table_names = c(I = "FAAt407"),
  start_date = "1940-01-01",
  end_date = "2023-12-31"
)
}
\arguments{
\item{table_names}{Name of tables to download from BEA.}

\item{start_date}{Starting date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}

\item{end_date}{Ending date in any format recognized by \link[lubridate:ymd]{lubridate::ymd}}
}
\description{
\code{get_bea()} downloads BEA data.
}
